Adding more logic to the Task  contract
Created a Task struct and a Priotization stuct

Updated logic for staking in the Validation of the ERC20

Added calculateWinners() method to see which vote won

Added an updatedLosersStake() method to change the staking of the losers and also update the prize pool for the winners.

What needs to be done next:
Unstake to let losers get their rewards and winners get their stake.
settle the task.